Fix warnings

+++ tools/viewer/viewer.cc
@@ -44,7 +44,7 @@ int main(int argc, char* argv[]) {
 
   fs.seekg(0, std::ios_base::end);
   auto end = fs.tellg();
-  if (db_hdr.end_of_file > end) {
+  if (db_hdr.end_of_file > static_cast<uint64_t>(end)) {
     std::cout << "Corrupted: file size < end of file marker!\n";
     return 1;
   }
